pyproject.toml
[project]
name = "vista-vid"
version = "0.1.0"
description = "Vista-VID Synthetic Dataset"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai==1.20.0",
    "python-dotenv==1.1.0",
    "moviepy==1.0.3",
    "pytube==15.0.0",
    "requests==2.31.0",
    "yt-dlp==2023.11.14",
]
